requirements.txt
streamlit
torch
transformers
openai
python-dotenv